projects/unit3/build-mcp-server/starter/pyproject.toml
[project]
name = "pr-agent"
version = "1.0.0"
description = "MCP server for PR template suggestions - Starter Code"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gitpython>=3.1.45",
    "mcp[cli]>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.21.0",
]
